+++ packages/gatsby-theme-carbon/src/components/FeedbackDialog/LaunchButton.js
@@ -1,17 +1,13 @@
 import React from 'react';
 import cx from 'classnames';
-import styles from './LaunchButton.module.scss';
+import { button } from './LaunchButton.module.scss';
 
-function LaunchButton({ onClick, visible, icon: RegularIcon }) {
-  const classNames = cx(styles.button, {
-    [styles.selected]: visible,
-  });
-
+function LaunchButton({ onClick, icon: RegularIcon }) {
   return (
     <button
       type="button"
       aria-haspopup
-      className={classNames}
+      className={button}
       onClick={onClick}
       aria-label="This launches a modal form to give website feedback.">
       <RegularIcon />